misc/100-discordia/src/requirements.txt
discord.py==3.0.1
